fixup! n-api,src: provide asynchronous cleanup hooks
